jwt token stored in the client is verified on the client everytime authContext mounts, so if token expires or is tampered with the user will be logged out of the application automatically

+++ api/sql/init.sql
@@ -10,7 +10,7 @@ CREATE TABLE contacts
     id SERIAL PRIMARY KEY,
     first_name VARCHAR NOT NULL,
     last_name VARCHAR,
-    phone_number INT NOT NULL,
+    phone_number BIGINT NOT NULL,
     company VARCHAR,
     email VARCHAR,
     address VARCHAR,
